feat: set cookies to be more secure with tls and httpOnly

+++ server/models/transactions.ts
@@ -1,12 +1,13 @@
 import {prisma} from './prisma'
 
-export async function createTransaction(date: string, description: string, amount: number, categoryId: number) {
+export async function createTransaction(date: string, description: string, amount: number, categoryId: number, userId: number | undefined) {
     return prisma.transactions.create({
         data: {
             amount,
             date,
             description: description.toLowerCase(),
-            categoryId
+            categoryId,
+            userId
         }
     });
 }
@@ -15,11 +16,11 @@ export async function updateTransaction(trxId: number, description: string, amou
     return prisma.transactions.update({where: {id: trxId}, data: { description, amount, date, categoryId }})
 }
 
-export async function deleteTransaction(trxId: number) {
-    return prisma.transactions.delete({where: {id: trxId}})
+export async function deleteTransaction(trxId: number, userId: number | undefined) {
+    return prisma.transactions.delete({where: {id: trxId, userId}})
 }
 
-export async function getTransactions(key: string, dateFilter: { start: string, end: string }) {
+export async function getTransactions(key: string, dateFilter: { start: string, end: string }, userId: number | undefined) {
     return prisma.transactions.findMany({
         where: {
             description: {
@@ -28,7 +29,8 @@ export async function getTransactions(key: string, dateFilter: { start: string, 
             date: {
                 gte: new Date(dateFilter.start),
                 lte: new Date(dateFilter.end),
-            }
+            },
+            userId
         },
         orderBy: {
             date: 'desc'
